filter in laptop tracking
